basic shmget code

+++ defs.h
@@ -191,7 +191,7 @@ void            clearpteu(pde_t *pgdir, char *uva);
 #define NELEM(x) (sizeof(x)/sizeof((x)[0]))
 
 //additional
-int shmget(char*, long, int);
+int shmget(char*, int, int);
 void* shmat(int, const void*, int);
 int shmdt(const void *);
 int shmctl(int, int, struct shmid_ds*);
